refs #38: refactor adaptor
