+++ internal/adaptor/repository/mock/client.go
