+++ internal/adaptor/repository/mysql/client_test.go
