+++ internal/adaptor/repository/mysql/event_test.go
