+++ internal/adaptor/repository/mysql/shcema/boil_queries.go
