+++ internal/adaptor/repository/mysql/shcema/boil_table_names.go
