+++ internal/adaptor/repository/mysql/shcema/boil_view_names.go
