+++ internal/adaptor/repository/mysql/shcema/device.go
