+++ internal/adaptor/transfer/mock/file.go
